refactor locations of front end pages add teams controller
